mb7420: define partitions, support bootloader console
Tested on a device that was active on Comcast up until a week ago, so
presumably this reflects the latest bootloader version.

I could not test or improve the firmware console profile, as that
console is disabled on my device (running firmware 5.7.1.31), even over
serial with dynnv.bfc.serial_console_mode="rw".