gwsettings: put a lower bound on segment size (#71)
On a Motorola MB7420 running firmware 5.7.1.31, nv files written by
bcm2cfg currently fail to load, with this in the log:

  Reading Dynamic settings from non-vol...
  NonVolDeviceDriver::ReadTrueSegmentSizeBytes:  WARNING - Segment Size in segment control word not valid; returning 0!
  NonVolDeviceDriver::NonVolDriverRead:  ERROR - Failed to query the segment size (returned 0 bytes)!  There is no valid data to read.
  [00:00:00 01/01/1970] [tStartup] BcmNonVolDeviceDriverBridge::ReadImpl:  (NonVol Device) ERROR - Failed to read the length from the settings data!
  [00:00:00 01/01/1970] [tStartup] BcmNonVolDeviceDriverBridge::Read:  (NonVol Device) ERROR - Failed to read the buffer from the device!  Resetting the entire section to defaults.
  [00:00:00 01/01/1970] [tStartup] CmSnmpNonVolSettings::ResetDefaults:  (CM SNMP NonVol Settings) WARNING - Resetting Dynamic Settings!

Ghidra disassembles the relevant parsing code as:

  if ((*p_part_size / denom_arr[param_1] < segment_size) || (segment_size < *p_part_size >> 4)) {
    printf("NonVolDeviceDriver::%s:  %s - Segment Size in segment control word not valid; returning 0!\n"
           ,"ReadTrueSegmentSizeBytes","WARNING");
    segment_size = 0;
  }

The second check, "segment_size < *p_part_size >> 4", is the one that's
failing. A right shift by 4 is equivalent to an integer division by 16,
so effectively this says that segment sizes smaller than 1/16 of the
partition are not valid.

I don't know if this check is present on all devices, but it shouldn't
hurt to work around it unconditionally. That's what this patch does.
Also fix a bug in the later calculations that was exposed by this
change.